# Optimization: filter project object before using regex