in /project/ views.

This significantly improves (~40-70% db speedup) the query made to find ProjectResources.

Timings:

/projects/firefox/browser/

NEW: 8 queries run, total 0.094 seconds
OLD: 8 queries run, total 0.138 seconds

/projects/firefox/browser/chrome/browser/customizableui/customizableWidgets.properties.po

NEW: 8 queries run, total 0.286 seconds
OLD: 8 queries run, total 0.722 seconds