@julen 2016-01-22
Thank you, @phlax, having numbers really helps.

Here are my results for a project with ~2.5M words, in the avg. of 5 runs always starting with an empty cache.

 | /projects/website2_evernote/ | /projects/website2_evernote/business/certified-consultants/inc/right-column.php.po
---- | ---- | ----
master | 18 queries, 0.0682s | 18 queries, 0.1348s
PR/4413 | 18 queries, 0.079s | 18 queries, 0.1034s

There seems to be a small regression when accessing the root of a project.

With the cache already filled, also the avg. of 5 runs:

 | /projects/website2_evernote/ | /projects/website2_evernote/business/certified-consultants/inc/right-column.php.po
---- | ---- | ----
master | 9 queries, 0.0034s | 9 queries, 0.072s
PR/4413 | 9 queries, 0.0028s | 9 queries, 0.0396s

The improvement doesn't look as visible as in your tests, however I assume your difference comes from the fact of having a project almost 4x bigger. I'm curious to know in which conditions you have measured your numbers though.