@phlax 2016-01-22
hi @julen im thinking that the benefit here will depend alot on the db.

ie  - if the first filter reduces the dataset a lot then there will be more benefit.

in terms of test infrastructure - im happy to PM you the details/provide access if you ping me on gitter