@phlax 2016-01-22
> There seems to be a small regression when accessing the root of a project.

Ive seen similar results in other tests but not in this particular case.

Im not quite sure why this would be my thought is that the regex is the more expensive query and therefore would always be better last, but in a very small minority of cases this doesnt seem to be true.

Mostly though by adding this query pattern (or similar) ive seen quite significant performance improvements.