@phlax 2016-01-22
@julen - just realising the point about the root of the /projects/ dir - we probs need to not filter __endwith in this case as its an empty string.