@phlax 2016-01-23
@julen i have investigated this further. This commit makes *no* changes to either the sql or python for views of the sort /projects/PROJECT/ - as the method does/did return the project without building any resources if dir_path and filename are empty - https://github.com/phlax/pootle/blob/view_optimizations_regex/pootle/apps/pootle_project/views.py#L64

so any differences in timings on these views are not related to this PR afaict