@ta2-1 2016-01-25
+1 to https://github.com/translate/pootle/pull/4413#issuecomment-174157751.  There are no changes for `/projects/PROJECT_NAME` pages.
And there are obvious improvements since we check less rows by `pootle_path` via regex. 
lgtm.