@julen 2016-01-25
> This commit makes no changes to either the sql or python for views of the sort /projects/PROJECT/

Thanks for the explanation, @phlax!